+++ src/main/java/com/codingMate/answer/repository/AnswerWriteRepository.java
@@ -1,30 +1,22 @@
 package com.codingMate.answer.repository;
 
 import com.codingMate.answer.domain.Answer;
-import com.codingMate.answer.domain.vo.LanguageType;
-import com.codingMate.answer.dto.response.QAnswerListResponse;
-import com.codingMate.programmer.domain.Programmer;
 import com.codingMate.answer.dto.request.AnswerCreateRequest;
 import com.codingMate.answer.dto.request.AnswerUpdateRequest;
-import com.codingMate.answer.dto.response.AnswerListResponse;
-import com.querydsl.core.types.dsl.Wildcard;
+import com.codingMate.programmer.domain.Programmer;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
 import jakarta.validation.constraints.NotNull;
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 import static com.codingMate.answer.domain.QAnswer.answer;
 
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class CustomAnswerRepository {
+@Repository
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
+public class AnswerWriteRepository {
     private final DefaultAnswerRepository answerRepository;
     private final JPAQueryFactory queryFactory;
     private final EntityManager em;
@@ -35,45 +27,6 @@ public class CustomAnswerRepository {
         entity.setProgrammer(programmer);
 
         return answerRepository.save(entity);
-    }
-
-    @Transactional(readOnly = true)
-    public Answer read(Long answerId) {
-        log.info("read({})", answerId);
-        return queryFactory.selectFrom(answer)
-                .where(answer.id.eq(answerId))
-                .join(answer.programmer)
-                .fetchOne();
-    }
-
-    @Transactional(readOnly = true)
-    public long wroteAnswerByProgrammer(Long programmerId) {
-        return queryFactory.select(Wildcard.count)
-                .from(answer)
-                .where(answer.programmer.id.eq(programmerId))
-                .fetchCount();
-    }
-
-    @Transactional(readOnly = true)
-    public List<AnswerListResponse> readAll(LanguageType languageType, Long backjoonId) {
-        log.info("readAll()");
-        return queryFactory.select(new QAnswerListResponse(answer.id, answer.backJoonId, answer.title, answer.programmer.name.name, answer.languageType))
-                .from(answer)
-                .where(languageType != null ? answer.languageType.eq(languageType) : null)
-                .where(backjoonId != null ? answer.backJoonId.eq(backjoonId) : null)
-                .join(answer.programmer)
-                .fetch();
-    }
-
-    @Transactional(readOnly = true)
-    public List<AnswerListResponse> readAllByProgrammerId(LanguageType language, Long backjoonId, Long programmerId) {
-        return queryFactory.select(new QAnswerListResponse(answer.id, answer.backJoonId, answer.title, answer.programmer.name.name, answer.languageType))
-                .from(answer)
-                .where(answer.programmer.id.eq(programmerId))
-                .where(language != null ? answer.languageType.eq(language) : null)
-                .where(backjoonId != null ? answer.backJoonId.eq(backjoonId) : null)
-                .join(answer.programmer)
-                .fetch();
     }
 
     @Transactional
