+++ src/main/java/com/codingMate/programmer/repository/ProgrammerWriteRepository.java
@@ -1,39 +1,31 @@
 package com.codingMate.programmer.repository;
 
-import com.codingMate.programmer.domain.Programmer;
 import com.codingMate.auth.domain.Authority;
+import com.codingMate.programmer.domain.Programmer;
 import com.codingMate.programmer.domain.vo.Email;
 import com.codingMate.programmer.domain.vo.Name;
 import com.codingMate.programmer.dto.request.ProgrammerCreateRequest;
 import com.codingMate.programmer.dto.request.ProgrammerUpdateRequest;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import jakarta.persistence.EntityManager;
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
-import static com.codingMate.auth.domain.QAuthority.authority;
 import static com.codingMate.programmer.domain.QProgrammer.programmer;
 
-
-@Slf4j
-@Service
-@RequiredArgsConstructor
-public class CustomProgrammerRepository {
+@Repository
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
+public class ProgrammerWriteRepository {
     private final DefaultProgrammerRepository programmerRepository;
     private final JPAQueryFactory queryFactory;
     private final EntityManager em;
-
-
     /**
      * @implSpec loginId가 중복되는지 여부 확인 + Programmer 등록으로 쿼리가 두 번 나감
      * */
     @Transactional
     public Programmer create(ProgrammerCreateRequest dto) {
-        log.info("create({})", dto);
         Authority authority = Authority.builder()
                 .authorityName("ROLE_USER")
                 .build();
@@ -51,53 +43,6 @@ public class CustomProgrammerRepository {
         return programmerRepository.save(entity);
     }
 
-    /**
-     * @implSpec * exists를 사용하지 않음으로써 성능 개선
-     *           * limit 1을 사용해서 1개를 찾으면 쿼리를 중단하도록 함
-     * */
-    @Transactional(readOnly = true)
-    public boolean isExistLoginId(String loginId) {
-        log.info("isExistLoginId({})", loginId);
-        return queryFactory.selectOne()
-                .from(programmer)
-                .where(programmer.loginId.eq(loginId))
-                .limit(1)
-                .fetchFirst() != null;
-    }
-
-    @Transactional(readOnly = true)
-    public Programmer read(Long id) {
-        log.info("read({})", id);
-        return programmerRepository
-                .findById(id)
-                .orElse(null);
-    }
-
-    @Transactional(readOnly = true)
-    public Programmer readByLoginId(String loginId) {
-        log.info("readByLoginId({})", loginId);
-        return queryFactory.selectFrom(programmer)
-                .where(programmer.loginId.eq(loginId))
-                .leftJoin(programmer.authority, authority)
-                .fetchOne();
-    }
-
-    @Transactional(readOnly = true)
-    public List<Programmer> readAll() {
-        log.info("readAll()");
-        return programmerRepository
-                .findAll();
-    }
-
-    @Transactional(readOnly = true)
-    public Long readIdByUserName(String userName) {
-        log.info("readIdByUserName({})", userName);
-        return queryFactory.select(programmer.id)
-                .from(programmer)
-                .where(programmer.loginId.eq(userName))
-                .fetchOne();
-    }
-
     @Transactional
     public long update(Long programmerId, ProgrammerUpdateRequest dto) {
         return queryFactory.update(programmer)
@@ -111,7 +56,6 @@ public class CustomProgrammerRepository {
 
     @Transactional
     public boolean delete(Long id) {
-        log.info("remove({})", id);
         if (programmerRepository.existsById(id)) {
             programmerRepository.deleteById(id);
             return true;
